added select books button to bring up delete checkboxes
