+++ src/main/client/src/dashBoard/bookList/book/Book.js
@@ -7,10 +7,12 @@ export default class Book extends Component {
         return (
             <li>
                 <h3>{book.title}
+                {this.props.deleteMode && 
                     <label className="container">
                         <input type="checkbox" value={book.id} onClick={this.props.handleCheck}></input>
                         <span className="checkmark"></span>
                     </label>
+                }
                 </h3>
                 <p>
                     author: {book.author},
